Change SWIG wrapper build method and name to obspython
This is a large update that changes the way the SWIG wrapper
is incorporated into the project. It should now be possible to
update the wrapper or obs plugin without recompiling the other.

It also allows the wrapper to be imported normally into python,
however it will fail to import if you attempt to import it into
anything but the embedded python as symbols will be missing.

The majority of the source methods have been implemented,
but are untested.